requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
requests==2.31.0
selenium==4.12.0
webdriver-manager==4.0.0
